refactor: move background script and content script
